(refactor) Refactor chart-review component
