+++ packages/esm-patient-chart-app/src/view-components/dashboard-view.scss
